MAJOR - FIX : error with the case KAD + manual Tibial Torsion.
With manual tibial Torsion, the AJC fro KAD is computed with the chord method and the tibial torsion.

I found out the foot offset were computed with the proximal tibia !!

Therefore, I added the option "viconCompatible" in calibration filter so to clone vicon.
With the CGM11,  foot offset will be computed relative to the distal segment.